Atsushi Nemoto writes:
I found inappropriate data types are used in some places in networking
codes.

* tcp_seq is 32bit (u_long -> u_int32_t)
* in_addt_t should be used for internet address (unsigned long -> in_addr_t)
* socklen_t should be used for accept()

This is a patch against uclibc-0.9.21 (can be applied for current
CVS).  64bit platforms (sizeof(int)!=sizeof(long)) will need this.  I
believe this patch does not harm any 32bit platforms.

+++ include/netinet/ip_tcp.h
@@ -24,7 +24,7 @@
 #include <linux/socket.h>
 #include <sys/types.h>
 
-typedef	u_long	tcp_seq;
+typedef	u_int32_t	tcp_seq;
 /*
  * TCP header.
  * Per RFC 793, September, 1981.
